added getAll method in organizationController

+++ user-service/src/main/java/com/example/user/model/User.java
@@ -23,5 +23,5 @@ public class User {
     String provider;
 
     @Column("organization_id")
-    int organizationId;
+    Long organizationId;
 }
